+++ backend/src/Weather.Api/Modules/Forecast/Endpoints/Requests/GetWeatherForecastRequest.cs
@@ -2,7 +2,7 @@ namespace Weather.Api.Modules.Forecast.Endpoints.Requests;
 
 public class GetWeatherForecastRequest
 {
-    public int? Latitude { get; set; }
-    public int? Longitude { get; set; }
+    public double? Latitude { get; set; }
+    public double? Longitude { get; set; }
     public char? TemperatureUnit { get; set; }
 }
